🚚 Renamed unsafe_hodl tests
