+++ apps/contracts/strategies/unsafe_hodl/src/test.rs
@@ -62,4 +62,4 @@ impl<'a> HodlStrategyTest<'a> {
     // }
 }
 
-mod hodl;
+mod unsafe_hodl;
