+++ apps/contracts/strategies/unsafe_hodl/src/test/unsafe_hodl/deposit.rs
@@ -1,0 +1,24 @@
+use crate::test::create_hodl_strategy;
+use crate::test::HodlStrategyTest;
+use crate::test::StrategyError;
+// test deposit with negative amount
+#[test]
+fn deposit_with_negative_amount() {
+   assert!(true);
+}
+
+// check auth
+#[test]
+fn deposit_mock_auths() {
+    assert!(true);
+}
+
+#[test]
+fn deposit_and_withdrawal_flow() {
+    assert!(true);
+}
+
+#[test]
+fn deposit_from_a_withdrawal_from_b() {
+    assert!(true);
+}
