+++ apps/contracts/strategies/unsafe_hodl/src/test/unsafe_hodl/events.rs
@@ -1,0 +1,18 @@
+use crate::event::{DepositEvent, HarvestEvent, WithdrawEvent};
+use crate::test::{create_hodl_strategy, HodlStrategyTest};
+use soroban_sdk::{symbol_short, testutils::Events, vec, IntoVal};
+
+#[test]
+fn deposit() {
+    assert!(true);
+}
+
+#[test]
+fn withdraw() {
+    assert!(true);
+}
+
+#[test]
+fn harvest() {
+    assert!(true);
+}
