+++ apps/contracts/strategies/unsafe_hodl/src/test/unsafe_hodl/withdraw.rs
@@ -1,0 +1,8 @@
+use crate::test::create_hodl_strategy;
+use crate::test::HodlStrategyTest;
+use crate::test::StrategyError;
+
+#[test]
+fn withdraw() {
+    assert!(true);
+}
